refactor: tech debt cleanup (#213, #206)
#213 - Extract jwkToPem to shared module:
- Create apps/registry/src/lib/crypto/jwk.ts
- Export jwkToPem() and base64UrlDecode()
- Update /auth/+server.ts to import from shared module

#206 - Address review comments:
- generateId(): Consistent 21-char total length with prefix
- trimOrNull(): Explicit null/undefined handling (value == null)
- Add forbiddenError(403) to api-responses.ts

Closes #213
