+++ apps/vault/src/lib/server/utils/id.ts
@@ -6,9 +6,13 @@
 /**
  * Generate a unique ID for database records
  * @param prefix Optional prefix for namespaced IDs (e.g., 'org_', 'invite_')
- * @returns A 21-character alphanumeric ID, optionally prefixed
+ * @returns A 21-character alphanumeric ID (or prefix + remaining chars to total 21)
  */
 export function generateId(prefix?: string): string {
-	const id = crypto.randomUUID().replace(/-/g, '').slice(0, 21);
-	return prefix ? `${prefix}${id.slice(0, 16)}` : id;
+	const base = crypto.randomUUID().replace(/-/g, '');
+	if (!prefix) {
+		return base.slice(0, 21);
+	}
+	// With prefix: total length is 21 chars (prefix + remaining)
+	return `${prefix}${base.slice(0, 21 - prefix.length)}`;
 }
